+++ elfo-core/src/subscription.rs
@@ -16,8 +16,15 @@ impl SubscriptionManager {
         }
     }
 
-    pub(crate) fn add(&self, addr: Addr) {
-        self.subscribers.write().push(addr);
+    pub(crate) fn add(&self, addr: Addr) -> bool {
+        let mut subscribers = self.subscribers.write();
+
+        if subscribers.contains(&addr) {
+            return false;
+        }
+
+        subscribers.push(addr);
+        true
     }
 
     pub(crate) fn remove(&self, addr: Addr) {